split timeline_clip into small files
