+++ lib/views/widgets/timeline/painters/image_grid_painter.dart
@@ -1,0 +1,22 @@
+import 'package:flutter/widgets.dart';
+
+class ImageGridPainter extends CustomPainter {
+  final Color color;
+
+  ImageGridPainter({required this.color});
+
+  @override
+  void paint(Canvas canvas, Size size) {
+    if (size.width <= 0 || size.height <= 0) return;
+    final paint = Paint()
+      ..color = color
+      ..strokeWidth = 1
+      ..style = PaintingStyle.stroke;
+    canvas.drawLine(Offset(0, 0), Offset(size.width, size.height), paint);
+    canvas.drawLine(Offset(size.width, 0), Offset(0, size.height), paint);
+    canvas.drawRect(Rect.fromLTWH(4, 4, size.width-8, size.height-8), paint);
+  }
+
+  @override
+  bool shouldRepaint(covariant CustomPainter oldDelegate) => false;
+}
