+++ lib/views/widgets/timeline/painters/video_frames_painter.dart
@@ -1,0 +1,29 @@
+import 'package:flutter/widgets.dart';
+
+class VideoFramesPainter extends CustomPainter {
+  final Color color;
+
+  VideoFramesPainter({required this.color});
+
+  @override
+  void paint(Canvas canvas, Size size) {
+    if (size.width <= 0 || size.height <= 0) return;
+    final paint = Paint()
+      ..color = color
+      ..strokeWidth = 1
+      ..style = PaintingStyle.stroke;
+    final cellWidth = size.width / 4;
+    final cellHeight = size.height / 3;
+    for (int i = 1; i < 4; i++) {
+      final x = cellWidth * i;
+      canvas.drawLine(Offset(x, 0), Offset(x, size.height), paint);
+    }
+    for (int i = 1; i < 3; i++) {
+      final y = cellHeight * i;
+      canvas.drawLine(Offset(0, y), Offset(size.width, y), paint);
+    }
+  }
+
+  @override
+  bool shouldRepaint(covariant CustomPainter oldDelegate) => false;
+}
